refactor(library): extract route helpers into typed modules (issue #124 stage A)
Move the ~1200-line helper preamble of routes/library.ts into cohesive
modules (libraryMaintenance service; libraryAudioInfo, libraryCoverArt,
libraryRadioPredicates, libraryRouteSupport, librarySorting utils) and
type the sort maps with the generated Prisma OrderByWithRelationInput
types. Route registrations and handler bodies are unchanged; this is a
readability-only split staged per the issue's sequencing.
